refactor(core): migrate workspace runtime providers
