+++ src/crates/services/services-core/Cargo.toml
@@ -11,7 +11,9 @@ crate-type = ["rlib"]
 
 [dependencies]
 anyhow = { workspace = true, optional = true }
+async-trait = { workspace = true, optional = true }
 bitfun-core-types = { path = "../../contracts/core-types" }
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", optional = true }
 tokio = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -33,6 +35,7 @@ win32job = { workspace = true }
 [features]
 default = ["lsp"]
 lsp = ["dep:anyhow", "dep:notify", "dep:zip"]
+workspace-runtime = ["dep:anyhow", "dep:async-trait", "dep:bitfun-runtime-ports"]
 
 [dev-dependencies]
 tempfile = { workspace = true }
