+++ src/crates/services/services-integrations/tests/remote_ssh_disabled_contracts.rs
@@ -1,3 +1,5 @@
+#![cfg(all(feature = "remote-ssh", not(feature = "remote-ssh-concrete")))]
+
 use std::path::PathBuf;
 use std::sync::Arc;
